Add files via upload
Provided HTML code represents a Registration Form for a course. It includes various input fields such as text boxes for name, email, and phone number, a date picker for date of birth, and text area for the address. Users can select their gender, branch, and college year through radio buttons, dropdowns, and checkboxes. An ordered list highlights the skills to be learned (web design, graphics design, and UI/UX design). Additionally, the form incorporates a payment section with dropdowns for card type, fields for card details, and an expiration date. It concludes with submit and reset buttons, plus a course information link.