Add files via upload
The given HTML and CSS code demonstrates the use of a CSS Media Query to create a responsive design. The default styles set the body with a bisque background color and brown text color. 

A media query is defined using `@media only screen and (max-width: 600px)`, which applies specific styles when the screen width is 600 pixels or less. In this case, the background color changes to red, and the text color becomes white for smaller screens, ensuring better readability and adaptability for mobile devices. This approach enhances user experience across different screen sizes.