Add files via upload
The provided HTML and CSS code demonstrates the use of  CSS Flexbox to create a responsive layout. The flex-container div uses display: flex and flex-wrap: wrap properties to arrange its child elements (flex-item) in a flexible and responsive row, wrapping to the next line if necessary.

Each flex-item represents a letter styled with a 75x75px size, center-aligned text, and a consistent font size and weight. Items have a light green (rgb(221, 249, 182)) background, a black border, bold blue text, and even spacing using the line-height property. 

The layout efficiently displays the letters "SUHANI" in a structured and visually appealing manner, showcasing the power of Flexbox for modern web design.